Python3 support and use tempdir instead of working dir, fixes #15 and #20
Rewrote lines that didn't work in Python3, mostly by converting things
of unicode/bytes type back to strings, but also by changing the import
of ConfigParser which was renamed to configparser in Python3 (I used a
try-except to not change the old behaviour in Python2. Also switched to
generating files in tempdir.get_tempdir()/markdown-latex instead of the
working directory.