Fixed Python2 compatibility
There was a feature that I forgot to mention in a previous commit,
namely a timeout for the pdflatex process. Now I tested in Python2 and
it turned out that the timeout argument of subprocess.call is not
supported in Python2, so I had to write a proxy function to remove it if
the code is detected to run under Python2. Now everything works in both
Python versions.

I also split some lines which according to the Python style guide are
too long.